fixup! core: dt: take account type of interrupt in dt_get_irq()
like dt-bindings is specific by drivers, I could proposed you a second
possibility:
- Move the dt_get_irq in interrupt framework.
- Add a xlate function in itr_chip structure, which allow to translate
  the specific properties of driver.
- This xlate function will be initialize by driver while init step.
- The dt_get_irq function find the interrupt property then call the
  dedicated xlate function of interrupt driver
