+++ core/kernel/interrupt.c
@@ -3,8 +3,10 @@
  * Copyright (c) 2016-2019, Linaro Limited
  */
 
+#include <kernel/dt.h>
 #include <kernel/interrupt.h>
 #include <kernel/panic.h>
+#include <libfdt.h>
 #include <stdlib.h>
 #include <trace.h>
 #include <assert.h>
@@ -25,6 +27,24 @@ void itr_init(struct itr_chip *chip)
 {
 	itr_chip = chip;
 }
+
+#ifdef CFG_DT
+int dt_get_irq(const void *fdt, int node)
+{
+	const uint32_t *prop = NULL;
+	int len = 0;
+	int it_num = DT_INFO_INVALID_INTERRUPT;
+
+	if (!itr_chip || !itr_chip->xlate)
+		return it_num;
+
+	prop = fdt_getprop(fdt, node, "interrupts", &len);
+	if (!prop)
+		return it_num;
+
+	return itr_chip->xlate(prop, len);
+}
+#endif
 
 void itr_handle(size_t it)
 {
